Implement validation check for nondeterministic multi-outputs per discussion on #13
This commit adds a validator for uml.Activity objects that checks to see whether any of the Activity's ObjectFlow edges share sources that are not intended to handle parallel paths (i.e., ForkNode, DecisionNode).
If it finds such pairs, it warns that they may cause unexpected non-deterministic flow.
This commit also adds a test to make sure that the validator is behavior correctly and fixes a few bugs caused by typos