Address #13 through implicit injection of ForkNodes
When an ObjectFlow is added using the Activity.use_value helper function, the code now automatically checks to see if adding this edge would create nondeterministic flow. If so, it injects/uses a ForkNode in order to allow the value to be used in multiple locations.

The commit also incorporates a couple of updates based on fixing typo bugs.