Replace compositionalProperty with isCompositionalProperty.
Previously all compositional properties were made as sub properties of
compositionalProperty. But this meant that compositionalProperty
contained all the pairs in all of its sub-properties -- an extensional
property -- not subclassing -- a behavioral property.

The only way I could think of to get the intended behavior in a
dialect of OWL that tools would support (i.e., *not* OWL Full) was to
replace it with an `owl:AnnotationProperty` that would take boolean
values (and would have to be treated as defaulting to false).