Pull LUDOX protocol out of tests.
Take the building of the LUDOX protocol object and file out of the body of the test, and move it to a python script in the examples/ directory.  The test now will load and run this procedure, making it reusable.
The top level `ludox_protocol` procedure is decomposed into sub-procedures, to make it possible to demonstrate it step-by-step, the way @dbryce has done in his jupyter notebook.