Change type of EmptyContainer specification
Changed from `paml:Identified` to `xsd:string`.  This permits us to store an OWL DL query string here instead of just a class name.
We took this choice because if we used a new, constructed OWL class identifier, we don't have a good way of extracting the definition of that class from the RDF store.