Fix the malformed OWL restrictions.
Add Jupyter notebook with the code for finding and fixing ill-formed
owl:Restriction definitions.

Also add revised versions of the paml.ttl and uml.ttl files.

Finally, add pairs of .nt files that can be diffed to see what
happened in the repair.

I propose we could add my code for finding bad restrictions to the git
repo as a before-commit hook.  To do so, we would need to pull the
code out of the notebook.